fix(tests): address review observations — naming, filler tests, and structural issues
- test_team.py: rename test_create_new_method_works → test_create_new_increments_team_count
- test_agent_job.py: extract restore/migrate status-transition tests into
  TestSiteRestoreJobStatusTransitions and TestSiteMigrateJobStatusTransitions;
  remove set_single_value reset from test body (tearDown rollback handles it)
- test_press_webhook.py: merge TestValidateEndpointUrlFormatValid/Invalid into
  TestValidateEndpointUrlFormat; add assertIsNone to all 5 valid cases
- test_cluster.py: move VirtualMachine test classes to test_virtual_machine.py
  (TestVirtualMachineDataDiskSnapshot, VolumeHelpers, ServerRouting, StatusMaps)
- test_prometheus_alert_rule.py: split generic TestPrometheusAlertRule into
  TestPrometheusAlertRuleRendering (get_rule/get_route) and
  TestPrometheusAlertRuleValidation (validate guards)
- test_agent_request_failure.py: remove _now() noise helper; use datetime.now() inline
